pyproject.toml
[project]
name = "faturas-whatsapp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabriel Alves", email = "gabriel52.alves@gmail.com"},
]
dependencies = [
    "selenium>=4.27.1",
    "requests>=2.32.3",
    "pymupdf>=1.25.1",
    "python-dotenv>=1.0.1",
    "pillow>=11.1.0",
    "pyzbar>=0.1.9",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
